// src/types/prayer.ts
export interface PrayerHeroData {
  heading: string
  scriptureQuote: string
  scriptureReference: string
  description: string
  image: string
}

export interface PrayerTeamData {
  heading: string
  description: string
  avatarImages: string[]
  additionalCount: number
}

export interface IntercedeCardData {
  heading: string
  description: string
  linkLabel: string
  linkHref: string
}